+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Versioning](http://semver.org/spec/v2.0.0.html).
   - [Jason](https://hex.pm/packages/jason) is now used for JSON encoding. The
     JSON library can no longer be injected via configuration.
   - [`msgpax`](https://hex.pm/packages/msgpax) 1.x is no longer supported
+  - Logs are now sent in batches of 1000 instead of 5000 to comply with the
+  Timber library specification
 
 ### Removed